schema:additionalType is a property - not a value type
I noticed this while helping Elizabeth Waldorf (Tracelink) with some SHACL validation issues.  One of the glitches I noticed was that "organizationType" was not only being mapped to schema:additionalType (via "@id" : "schema:additionalType" ) - that's OK; what is not OK is casting a string value such as "Wholesaler" to "Wholesaler"^^schema:additionalType when schema:additionalType is not even a data type in the way that xsd:date is.